Fix the Windows build.
The build is now working again on windows.  I've stopped using
vcpkg as it doesn't allow me to build the cpprestsdk the way
I want (without websocketpp)

Document the build instructions so I don't forget.

TODO: Make a variable file or something to contain all the flags
needed to point to libs and include dirs