Change REST api to return raw pointers.
The parallel patterns library can't handle move only types.
Instead of forcing the user to take a std::shared_ptr or a copy
of a large object, I've changed the API to return a raw ptr.

Probably not ideal but I don't want to force the usage of a
std::shared_ptr on the user of the API.

The raw ptr can be easily converted into a unique or shared
pointer if desired by the user.