Add a cross-platform porting wrapper.
I'm not really happy with having the porting code floating about
in "porting.h" and random .cpp files.

I've introduced a platform library to attempt to abstract away
platform specefic operations such as the windows UTF8/16 nonsense
as well as things like file IO and whatever else we need to deal
with.

This is an early WIP.  Error handling is almost non-existent and
will need to be handled.