Reorder the JSON reformat statements so it's a bit easier to read.
I find this makes it a bit clearer about what is going on.

Also, handle the case where we get a parse error, although it's
probably pretty unlikely to happen.

The linux truncate command doesn't rewind the file pointer to
where you truncated the file, so do that manually before the
truncate.