Regenerate parsers to use the new string/integer routines.
I've given the parser generator the ability to read numbers in as
strings.  This is useful for when we want to read in currency values.
Doubles aren't suitable for currency as they can cause precision errors.

All the parsers have changed somewhat as the order of methods being
written out the file have changed somewhat, so there's a bit of
collateral damage to the commit history.

The only parser that currently uses the as_string feature is the
market_get parser.